Reverted back the change of the `database` structure.
