+++ fuel-core/src/database/coin.rs
