+++ fuel-core/src/database/contracts.rs
