+++ fuel-core/src/database/delegates_index.rs
