+++ fuel-core/src/database/staking_diffs.rs
