+++ fuel-core/src/database/state.rs
