+++ fuel-core/src/database/storage.rs
@@ -11,19 +11,6 @@ use fuel_core_interfaces::{
 use fuel_crypto::fuel_types::Bytes32;
 use fuel_txpool::types::ContractId;
 
-pub mod balances;
-pub mod block;
-pub mod code_root;
-pub mod coin;
-pub mod contracts;
-pub mod delegates_index;
-pub mod message;
-pub mod receipts;
-pub mod staking_diffs;
-pub mod state;
-pub mod transaction;
-pub mod validator_set;
-
 /// The table of blocks generated by Fuels validators.
 /// Right now, we have only that type of block, but we will support others in the future.
 pub struct FuelBlocks;
